+++ cfn-launcher.sh
@@ -1,10 +1,10 @@
 #!/bin/bash
 
-# !! WIP !!
-
 # Static Variables
 successful=false
-script_version=1.3.0-20161105
+delete_successful=false
+triggered_delete=false
+script_version=1.4.0
 # unset stack_name
 # read -p "Enter Stack Name: " stack_name
 
@@ -14,9 +14,9 @@ helpmessage="
 |||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
 -----------------------------------------------------------------------------------------------------------------------
 AUTHOR:       Levon Becker
-PURPOSE:      Create or Update CloudFormation Stack with CloudFormation Template.
+PURPOSE:      Create, Update or Delete CloudFormation Stack.
 VERSION:      $script_version
-DESCRIPTION:  This script uses the AWS CLI and BASH to create or update a CloudFormation Stack.
+DESCRIPTION:  This script uses the AWS CLI and BASH to create, update or delete a CloudFormation Stack.
               It uses the AWS CLI to push the CloudFormation Template to AWS.
               Then loops over and over checking status of the stack.
 -----------------------------------------------------------------------------------------------------------------------
@@ -31,10 +31,12 @@ capabilityiam: true
 capabilitynamediam: false
 deletecreatefailures: false
 uses3template: true
+nolog: false
 logfile: /path/to/where/you/want/logs/cfnl-awsaccount-region-env-stack.log
 verbose: true
 waittime: 5
 maxwaits: 180
+noheader: false
 -----------------------------------------------------------------------------------------------------------------------
 EXAMPLES
 -----------------------------------------------------------------------------------------------------------------------
@@ -43,6 +45,9 @@ $0 -c /path/to/cfnl/configs/awsaccount-region-env-stack-cfnlconfig.yml
 
 Update Stack
 $0 -u -c /path/to/cfnl/configs/awsaccount-region-env-stack-cfnlconfig.yml
+
+Delete Stack
+$0 -d -c /path/to/cfnl/configs/awsaccount-region-env-stack-cfnlconfig.yml
 "
     echo "$helpmessage";
 }
@@ -58,15 +63,17 @@ usage: $0 [-u] -c ./config_file.yml
 
 -c Config File           :  YAML Script Config File Full Path (Required)
 -u Update Stack          :  Triggers Update Operation (Default is Create Stack)
+-d Update Stack          :  Triggers Deletion of Stack
 -h Help                  :  Displays Help Information
 "
     echo "$usagemessage";
 }
 
-while getopts "c:uvh" opts; do
+while getopts "c:udvh" opts; do
     case $opts in
         c ) config_file_path=$OPTARG;;
         u ) update=true;;
+        d ) delete=true;;
         v ) version_message; exit 0;;
         h ) help_message; exit 0;;
     esac
@@ -103,12 +110,11 @@ function parse_yaml() {
 
 function message() {
     DATETIME=$(date +%Y-%m-%d_%H:%M:%S)
-    echo "[$DATETIME] $*" | tee -a ${yaml_logfile}
-}
-# TODO: Combine these two functions and use arg to switch modes
-function message_nofile() {
-    DATETIME=$(date +%Y-%m-%d_%H:%M:%S)
-    echo "[$DATETIME] $*"
+    if [ "$yaml_nolog" == "true" ]; then
+        echo "[$DATETIME] $*"
+    else
+        echo "[$DATETIME] $*" | tee -a ${yaml_logfile}
+    fi
 }
 
 function show_header {
@@ -123,10 +129,11 @@ function show_header {
 |||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
 -----------------------------------------------------------------------------------------------------------------------
 CloudFormation Launcher
-$script_version
+v$script_version
 -----------------------------------------------------------------------------------------------------------------------
 PARAMETERS
 -----------------------------------------------------------------------------------------------------------------------
+ACTION:               $ACTION
 STACK NAME:           $yaml_stackname
 PROFILE:              $yaml_profilename
 TEMPLATE:             $TEMPLATE
@@ -134,34 +141,38 @@ PARAMETERS FILE:      $yaml_parametersfilepath
 CAPABILITY IAM:       $yaml_capabilityiam
 CAPABILITY NAMED IAM: $yaml_capabilitynamediam
 TASK TYPE:            $task_type
+NO LOG:               $yaml_nolog
 LOG FILE:             $yaml_logfile
 VERBOSE:              $yaml_verbose
 LAUNCHER CONFIG:      $config_file_path
+DELETE ON FAILURE:    $yaml_deletecreatefailures
 WAIT TIME (Sec):      $yaml_waittime
 MAX WAITS (Loops):    $yaml_maxwaits
 -----------------------------------------------------------------------------------------------------------------------
   "
-	echo "$HEADER" | tee -a ${yaml_logfile};
+	if [ "$yaml_noheader" == "true" ]; then
+	    message ''
+	else
+		message "$HEADER"
+	fi
 }
 
 function exit_check {
-	if [ $1 -eq 0 ]
-	then
-		message "REPORT: Successfully $2" | tee -a ${yaml_logfile}
-	else
-		message "ERROR:  Exit Code $1 for $2" | tee -a ${yaml_logfile}
-		exit $1
-	fi
-}
-# TODO: Combine these two functions and have arg switch mode
-function exit_check_nolog {
-	if [ $1 -eq 0 ]
-	then
-		message_nofile "REPORT: Successfully $2"
-	else
-		message_nofile "ERROR:  Exit Code $1 for $2"
-		exit $1
-	fi
+    if [ "$triggered_delete" == "true" ]; then
+        if [[ $1 -eq 0 || $1 -eq 255 ]]; then
+            message "REPORT: Successfully $2"
+        else
+            message "ERROR:  Exit Code $1 for $2"
+            exit $1
+        fi
+    else
+        if [ $1 -eq 0 ]; then
+            message "REPORT: Successfully $2"
+        else
+            message "ERROR:  Exit Code $1 for $2"
+            exit $1
+        fi
+    fi
 }
 
 function run_stack_command {
@@ -189,78 +200,105 @@ function run_stack_command {
     fi
     exit_check $? "Executed ${ACTION} Stack Command"
     message ''
+
+    monitor_stack_status
 }
 
 function delete_stack_command {
     message 'ACTION: Deleting Stack'
     aws cloudformation delete-stack --profile ${yaml_profilename} --stack-name ${yaml_stackname}
     exit_check $? "Executed Delete Stack Command"
+
+    monitor_delete_stack_status
 }
 
 function output_create_complete {
     # If Verbose True then Output all the Create Complete Events for Debugging
     if [ "$yaml_verbose" == "true" ]; then
-        aws cloudformation describe-stack-events --stack-name ${yaml_stackname} --query 'StackEvents[?ResourceStatus==`CREATE_COMPLETE`]' | tee -a ${yaml_logfile}
-        echo '' | tee -a ${yaml_logfile}
+        message "REPORT: CREATE COMPLETE EVENTS..."
+        message ''
+        create_complete=$(aws cloudformation describe-stack-events --profile ${yaml_profilename} --stack-name ${yaml_stackname} --output json --query 'StackEvents[?ResourceStatus==`CREATE_COMPLETE`]')
+        message "$create_complete"
+        message ''
     fi
 }
 
 function output_create_failed {
     # Output all the Create Failed Events for Debugging
-    aws cloudformation describe-stack-events --stack-name ${yaml_stackname} --query 'StackEvents[?ResourceStatus==`CREATE_FAILED`]' | tee -a ${yaml_logfile}
+    message "REPORT: CREATE FAILED EVENTS..."
+    message ''
+    create_failed=$(aws cloudformation describe-stack-events --profile ${yaml_profilename} --stack-name ${yaml_stackname} --output json --query 'StackEvents[?ResourceStatus==`CREATE_FAILED`]')
+    message "$create_failed"
+    message ''
+}
+
+function monitor_delete_stack_status {
+    triggered_delete=true
+    wait_time=${yaml_waittime}
+    max_waits=${yaml_maxwaits}
+    while :
+    do
+        STATUS=$(aws cloudformation describe-stacks --profile ${yaml_profilename} --stack-name "$yaml_stackname" --output text --query 'Stacks[*].StackStatus')
+        exit_code=$?
+        exit_check ${exit_code} "Executing Status Check"
+        message "REPORT: Status (${STATUS})"
+
+        if [[ "$STATUS" == "DELETE_IN_PROGRESS" && ${count} -lt ${max_waits} ]]; then
+            message "REPORT: Delete not complete!"
+            message "REPORT: Attempt $count of $max_waits."
+            message "Polling again in $wait_time seconds..."
+            message ''
+            sleep ${wait_time}
+            count=$(( count + 1 ))
+        elif [ ${exit_code} -eq 255 ]; then
+            delete_successful=true
+            message "REPORT: Stack Deleted ($yaml_stackname)"
+            break
+        else
+            message 'ERROR: The stack delete has failed.'
+            break
+        fi
+    done
 }
 
 function monitor_stack_status {
-    # Poll for Status
-    # wait_time 5, max_waits 180 = 15 minutes
-
-    if [ "$task_type" == "create-stack" ]; then
-        ACTION=CREATE
-    elif [ "$task_type" == "update-stack" ]; then
-        ACTION=UPDATE
-    else
-        ACTION=CREATE
-    fi
-
     wait_time=${yaml_waittime}
     max_waits=${yaml_maxwaits}
-    count=1
-    delete_triggered=false
     while :
     do
-        STATUS=$(aws cloudformation describe-stacks --stack-name "$yaml_stackname" --output text --query 'Stacks[*].StackStatus')
+        STATUS=$(aws cloudformation describe-stacks --profile ${yaml_profilename} --stack-name "$yaml_stackname" --output text --query 'Stacks[*].StackStatus')
         exit_check $? "Executing Status Check"
         message "REPORT: Status (${STATUS})"
 
         if [[ "$STATUS" == "${ACTION}_IN_PROGRESS" && ${count} -lt ${max_waits} ]]; then
-            message "REPORT: ${ACTION} stack is not complete!"
+            message "REPORT: $ACTION stack is not complete!"
             message "REPORT: Attempt $count of $max_waits."
-            message "REPORT: Polling again in ${wait_time} seconds..."
-            echo '' | tee -a ${yaml_logfile}
+            message "REPORT: Polling again in $wait_time seconds..."
+            message ''
             sleep ${wait_time}
             count=$(( count + 1 ))
         elif [ "$STATUS" == "${ACTION}_COMPLETE" ]; then
-            message "REPORT: ${ACTION} Completed!"
+            message "REPORT: $ACTION Completed!"
             successful=true
             break
         elif [ "$STATUS" == "${ACTION}_FAILED" ]; then
-            message "ERROR:  ${ACTION} Failed!"
-            successful=false
+            message "ERROR:  $ACTION Failed!"
         elif [ "$STATUS" == "UPDATE_COMPLETE_CLEANUP_IN_PROGRESS" ]; then
             message 'REPORT: Cleanup in Progress'
             message "REPORT: Attempt $count of $max_waits."
-            message "REPORT: Polling again in ${wait_time} seconds..."
-            echo '' | tee -a ${yaml_logfile}
+            message "REPORT: Polling again in $wait_time seconds..."
+            message ''
             sleep ${wait_time}
             count=$(( count + 1 ))
         elif [ "$STATUS" == "ROLLBACK_IN_PROGRESS" ]; then
             # If Delete Stack on failures when Creating is True then Delete the Stack after grabbing Events
             if [[ "$task_type" == "create-stack" && "$yaml_deletecreatefailures" == "true" ]]; then
                 message 'ERROR:  Failed and Rolling Back!'
+                message ''
                 output_create_complete
                 output_create_failed
                 delete_stack_command
-                successful=false
+                break
             # Else Fetch Events, but do not Delete the Stack
             else
                 output_create_complete
@@ -268,58 +306,67 @@ function monitor_stack_status {
                 message 'ERROR:  Failed and Rolling Back!'
                 message "REPORT: Rollback not complete!"
                 message "REPORT: Attempt $count of $max_waits."
-                message "Polling again in ${wait_time} seconds..."
-                echo '' | tee -a ${yaml_logfile}
-                sleep $wait_time
+                message "Polling again in $wait_time seconds..."
+                message ''
+                sleep ${wait_time}
                 count=$(( count + 1 ))
-                successful=false
             fi
         elif [ "$STATUS" == "DELETE_IN_PROGRESS" ]; then
-            message "REPORT: Delete not complete!"
-            message "REPORT: Attempt ${count} of ${max_waits}."
-            message "Polling again in ${wait_time} seconds..."
-            echo '' | tee -a ${yaml_logfile}
-            sleep ${wait_time}
-            count=$(( count + 1 ))
-            successful=false
+            monitor_delete_stack_status
             break
         elif [ "$STATUS" == "ROLLBACK_COMPLETE" ]; then
             message "REPORT: Rollback complete!"
-            echo '' | tee -a ${yaml_logfile}
-            successful=false
+            message ''
             break
         else
             message 'ERROR: The stack has not create or update has failed.'
-            successful=false
+            message ''
             break
         fi
     done
 }
 
-# Start
-#get_args
-#validate_args
+# Start Time
 start_time=$(date +%s)
 # Read Yaml Properties File
 eval $(parse_yaml ${config_file_path} "yaml_")
 #set | grep yaml_
-run_stack_command
-monitor_stack_status
+count=1
+if [ "$delete" == "true" ]; then
+    ACTION=DELETE
+    show_header
+    delete_stack_command
+else
+    if [ "$task_type" == "create-stack" ]; then
+        ACTION=CREATE
+    elif [ "$task_type" == "update-stack" ]; then
+        ACTION=UPDATE
+    else
+        ACTION=CREATE
+    fi
+    run_stack_command
+fi
 
 # End Time
 end_time=$(date +%s)
 
 # Results
-echo '' | tee -a ${yaml_logfile}
+message ''
 message "ENDTIME: ($(date))"
 elapsed=$(( (${end_time} - ${start_time}) / 60 ))
 message "RUNTIME: ${elapsed} minutes"
-echo '' | tee -a ${yaml_logfile}
+message ''
 
-if [ "$successful" == "true" ]; then
-  message "REPORT: SUCCESS!"
-  exit 0
+if [[ "$delete_successful" == "true" && "$ACTION" == "DELETE" ]]; then
+    message "REPORT: DELETE SUCCESS!"
+    message ''
+    exit 0
+elif [ "$successful" == "true" ]; then
+    message "REPORT: $ACTION SUCCESS!"
+    message ''
+    exit 0
 else
-  message "ERROR: FAILED!"
-  exit $1
+    message "ERROR: $ACTION FAILED!"
+    message ''
+  exit 1
 fi